src/clear_db_daemon/requirements.txt
python-dotenv
pymongo
pytz